README improvements - bootstrap finished
